+++ dms-core/src/main/kotlin/team/aliens/dms/domain/user/service/CheckUserService.kt
@@ -1,0 +1,7 @@
+package team.aliens.dms.domain.user.service
+
+interface CheckUserService {
+    fun checkUserNotExistsByEmail(email: String)
+    fun checkUserExistsByEmail(email: String)
+    fun checkUserNotExistsByAccountId(accountId: String)
+}
