+++ dms-core/src/main/kotlin/team/aliens/dms/domain/user/service/GetUserService.kt
@@ -1,9 +1,9 @@
 package team.aliens.dms.domain.user.service
 
+import java.util.UUID
 import team.aliens.dms.domain.auth.model.Authority
 import team.aliens.dms.domain.student.model.Student
 import team.aliens.dms.domain.user.model.User
-import java.util.UUID
 
 interface GetUserService {
 
@@ -12,12 +12,6 @@ interface GetUserService {
     fun queryUserByAccountId(accountId: String): User
 
     fun queryUserBySchoolIdAndAuthority(schoolId: UUID, authority: Authority): User
-
-    fun checkUserNotExistsByEmail(email: String)
-
-    fun checkUserExistsByEmail(email: String)
-
-    fun checkUserNotExistsByAccountId(accountId: String)
 
     fun getCurrentUser(): User
